util/linuxfw,wgengine/router: export some of the functionality
So that it can be used by iptables/nftables in containerboot

Also exposes List method via iptables interface- to be used in tests.
