+++ util/linuxfw/linuxfw.go
@@ -131,7 +131,7 @@ func errCode(err error) int {
 // missing.  It does not check that IPv6 is currently functional or
 // that there's a global address, just that the system would support
 // IPv6 if it were on an IPv6 network.
-func checkIPv6(logf logger.Logf) error {
+func CheckIPv6(logf logger.Logf) error {
 	_, err := os.Stat("/proc/sys/net/ipv6")
 	if os.IsNotExist(err) {
 		return err
@@ -176,7 +176,7 @@ func checkIPv6(logf logger.Logf) error {
 // The nat table was added after the initial release of ipv6
 // netfilter, so some older distros ship a kernel that can't NAT IPv6
 // traffic.
-func checkSupportsV6NAT() bool {
+func CheckSupportsV6NAT() bool {
 	bs, err := os.ReadFile("/proc/net/ip6_tables_names")
 	if err != nil {
 		// Can't read the file. Assume SNAT works.
